Move IdGenerator to model.entry as it is inherently tied to BibtexEntry.
