+++ src/main/java/net/sf/jabref/gui/ImportInspectionDialog.java
@@ -67,7 +67,7 @@ import net.sf.jabref.model.entry.BibtexEntry;
 import net.sf.jabref.bibtex.DuplicateCheck;
 import net.sf.jabref.bibtex.comparator.FieldComparator;
 import net.sf.jabref.Globals;
-import net.sf.jabref.logic.id.IdGenerator;
+import net.sf.jabref.model.entry.IdGenerator;
 import net.sf.jabref.JabRefExecutorService;
 import net.sf.jabref.JabRefPreferences;
 import net.sf.jabref.MetaData;
