+++ src/main/java/net/sf/jabref/importer/ImportFormatReader.java
@@ -19,7 +19,7 @@ import java.io.*;
 import java.util.*;
 
 import net.sf.jabref.importer.fileformat.*;
-import net.sf.jabref.logic.id.IdGenerator;
+import net.sf.jabref.model.entry.IdGenerator;
 import net.sf.jabref.model.database.BibtexDatabase;
 import net.sf.jabref.model.entry.BibtexEntry;
 import org.apache.commons.logging.Log;
