+++ src/main/java/net/sf/jabref/importer/fetcher/CiteSeerXFetcher.java
@@ -17,7 +17,7 @@ package net.sf.jabref.importer.fetcher;
 
 import net.sf.jabref.importer.ImportInspector;
 import net.sf.jabref.importer.OutputPrinter;
-import net.sf.jabref.logic.id.IdGenerator;
+import net.sf.jabref.model.entry.IdGenerator;
 import net.sf.jabref.logic.net.URLDownload;
 import net.sf.jabref.logic.util.strings.NameListNormalizer;
 import net.sf.jabref.model.entry.BibtexEntry;
