+++ src/main/java/net/sf/jabref/importer/fileformat/JstorImporter.java
@@ -23,9 +23,8 @@ import java.util.ArrayList;
 
 import net.sf.jabref.importer.ImportFormatReader;
 import net.sf.jabref.importer.OutputPrinter;
-import net.sf.jabref.importer.fileformat.ImportFormat;
 import net.sf.jabref.model.entry.AuthorList;
-import net.sf.jabref.logic.id.IdGenerator;
+import net.sf.jabref.model.entry.IdGenerator;
 import net.sf.jabref.model.entry.BibtexEntry;
 import net.sf.jabref.model.entry.BibtexEntryType;
