+++ src/test/java/net/sf/jabref/groups/structure/ExplicitGroupTest.java
@@ -1,7 +1,7 @@
 package net.sf.jabref.groups.structure;
 
 import net.sf.jabref.*;
-import net.sf.jabref.logic.id.IdGenerator;
+import net.sf.jabref.model.entry.IdGenerator;
 import net.sf.jabref.model.entry.BibtexEntry;
 import net.sf.jabref.model.entry.BibtexEntryTypes;
 import org.junit.Before;
