+++ src/test/java/net/sf/jabref/importer/fileformat/OAI2ImportTest.java
@@ -4,7 +4,7 @@ import net.sf.jabref.importer.OAI2Handler;
 import net.sf.jabref.importer.fetcher.OAI2Fetcher;
 import net.sf.jabref.model.entry.BibtexEntry;
 import net.sf.jabref.model.entry.BibtexEntryType;
-import net.sf.jabref.logic.id.IdGenerator;
+import net.sf.jabref.model.entry.IdGenerator;
 
 import org.junit.Assert;
 import org.junit.Before;
