+++ src/test/java/net/sf/jabref/logic/search/rules/ContainBasedSearchRuleTest.java
@@ -2,7 +2,7 @@ package net.sf.jabref.logic.search.rules;
 
 import net.sf.jabref.*;
 
-import net.sf.jabref.logic.id.IdGenerator;
+import net.sf.jabref.model.entry.IdGenerator;
 import net.sf.jabref.model.entry.BibtexEntry;
 import net.sf.jabref.model.entry.BibtexEntryTypes;
 import org.junit.Assert;
