+++ src/test/java/net/sf/jabref/model/entry/IdGeneratorTest.java
@@ -13,9 +13,9 @@
     with this program; if not, write to the Free Software Foundation, Inc.,
     51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
-package net.sf.jabref.logic.util;
+package net.sf.jabref.model.entry;
 
-import net.sf.jabref.logic.id.IdGenerator;
+import net.sf.jabref.model.entry.IdGenerator;
 import org.junit.Assert;
 import org.junit.Test;
